add a short delay to bank ident fetch
